\name{aprof-package}
\alias{aprof-package}
\alias{aprof}
\docType{package}
\title{
 Amdahl's profiler, directed optimization made easy.
}
\description{
A R package meant to help evaluate whether and where to focus 
code optimization using Amdahl's law.

}
\details{
\tabular{ll}{
Package: \tab aprof\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-06-25\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Marco D. Visser

Maintainer: Who to complain to <yourfault@somewhere.net>
~~ The author and/or maintainer of the package ~~
}
\references{
Amdahl, Gene (1967). "Validity of the Single Processor Approach to Achieving Large-Scale Computing Capabilities".
AFIPS Conference Proceedings (30): 483–485.
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
