% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aprof.R
\name{is.aprof}
\alias{is.aprof}
\title{is.aprof}
\usage{
is.aprof(object)
}
\arguments{
\item{object}{Object to test}
}
\description{
Generic lower-level function to test whether an object
is an aprof object.
}

