\name{print.aprof}
\alias{print.aprof}
\title{Generic print method for aprof objects}
\usage{
  print.aprof(x, ...)
}
\arguments{
  \item{x}{An aprof object returned by the function
  \code{aprof}}

  \item{...}{Additional printing arguments}
}
\description{
  Function that makes a pretty, and returns some basic
  information
}

