\name{readOutput}
\alias{readOutput}
\title{readOutput}
\usage{
  readOutput(outputfilename = "Rprof.out")
}
\arguments{
  \item{outputfilename}{The file name (and path if not in
  the working directory) of a previously created profiling
  exercise.}
}
\description{
  Reads and organises output files created by the R
  profiler for further analysis. This program will be
  updated to generate a "aprof class" for default plotting
  and printing.
}
\author{
  Marco D. Visser
}

