\name{aprof}
\alias{aprof}
\title{Amdahl's profiler}
\usage{
  aprof(src = NULL, output = NULL, memoutput = NULL)
}
\arguments{
  \item{src}{The source code file name (and path if not in
  the working directory) of the program that as been
  profiled.}

  \item{output}{The file name (and path if not in the
  working directory) of a previously created profiling
  exercise.}

  \item{memprofile}{Optional. The file name (and path)}
}
\description{
  Amdahl's profiler, directed optimization.
}
\details{
  Creates an "aprof" object from the R-profiler's output
  and a source file.

  .. content for \details{} ..
}
\author{
  Marco D. Visser
}

