\name{print.aprof}
\alias{print.aprof}
\title{Generic print method for aprof objects}
\usage{
  print.aprof(aprofobject)
}
\arguments{
  \item{aprofobject}{An aprof object return by the function
  \code{aprof}}
}
\description{
  Function makes a pretty table, and returns some basic
  information
}

