% Generated by roxygen2 (4.0.0): do not edit by hand
\name{print.aprof}
\alias{print.aprof}
\title{Generic print method for aprof objects}
\usage{
\method{print}{aprof}(x, ...)
}
\arguments{
\item{x}{An aprof object returned by the
function \code{aprof}.}

\item{\dots}{Additional printing arguments. Unused.}
}
\description{
Function that makes a pretty table, and returns
some basic information
}

