% Generated by roxygen2 (4.0.1): do not edit by hand
\name{targetedSummary}
\alias{targetedSummary}
\title{targetedSummary}
\usage{
targetedSummary(target = NULL, aprofobject = NULL, findParent = FALSE,
  mem = FALSE)
}
\arguments{
\item{target}{The specific line of code to take a detailed look
at. This can be identfied using \code{summary.aprof}.}

\item{aprofobject}{object of class "aprof" returned by
the function \code{aprof}.}

\item{findParent}{Logical, should an attempt be made to find
the parent of a function call? E.g. "lm" would be a parent call of
"lm.fit" or "mean" a parent call of "mean.default".
Note that currently, the option only returns the most frequently
associated parent call when multiple unique parents exist.}

\item{mem}{Logical, should statistics be adapted to a memory
profile? This is only possible if the output from Rprofmem
was included in the aprof-object [Note: unavailable for this
release].}
}
\description{
Allows a detailed look into certain lines of code,
which have previously been identified as bottlenecks
in combination with a source file.
}
\author{
Marco D. Visser
}

